correction

+++ cpu_2d/compression/compression.hh
@@ -13,7 +13,7 @@ class Compression
 {
 public:
     virtual void debugCompression(T *fq, const size_t size) const = 0;
-    virtual void compress(T *fq_64, const size_t size, T_C **compressed_fq_64, size_t &compressedsize) const = 0;
+    virtual void compress(T *fq_64, const int32_t size, T_C **compressed_fq_64, size_t &compressedsize) const = 0;
     virtual void decompress(T_C *compressed_fq_64, const size_t size,
                             /*Out*/ T **uncompressed_fq_64, /*In Out*/size_t &uncompressedsize) const = 0;
     virtual void verifyCompression(const T *fq, const T *uncompressed_fq_64,
