+++ cpu_2d/vreduce.hpp
@@ -183,7 +183,7 @@ void vreduce(const function <void(T, long, T *, int)> &reduce,
 #endif
 
 #ifdef _COMPRESSION
-                schema.compress(send, (size_t)psizeTo, &compressed_fq, compressedsize);
+                schema.compress(send, psizeTo, &compressed_fq, compressedsize);
 
 #endif
 
@@ -240,7 +240,7 @@ void vreduce(const function <void(T, long, T *, int)> &reduce,
 #endif
 
 #ifdef _COMPRESSION
-                uncompressedsize = originalsize;
+                uncompressedsize = static_cast<size_t>(originalsize);
                 schema.decompress(temporal_recv_buff, psizeFrom, &uncompressed_fq, uncompressedsize);
 #endif
 
@@ -343,7 +343,7 @@ void vreduce(const function <void(T, long, T *, int)> &reduce,
 #endif
 
 #ifdef _COMPRESSION
-                uncompressedsize = originalsize;
+                uncompressedsize = static_cast<size_t>(originalsize);
                 schema.decompress(temporal_recv_buff, psizeFrom, &uncompressed_fq, uncompressedsize);
 #endif
 
@@ -425,7 +425,7 @@ void vreduce(const function <void(T, long, T *, int)> &reduce,
     uint32_t reversedSliceIDs;
     size_t csize = 0U;
 
-    schema.compress(send, (size_t)psizeTo, &compressed_fq, compressedsize);
+    schema.compress(send, psizeTo, &compressed_fq, compressedsize);
 
     int32_t * restrict composed_recv;
     int32_t * restrict composed_send;
